📑 Added PlaceOrder on COD for User

+++ internal/models/dish.go
@@ -11,7 +11,7 @@ type Dish struct {
 	RestaurantID uuid.UUID `json:"restaurantId" gorm:"type:uuid;foreignKey:restaurant.id;notNull"`
 	Name         string    `json:"dishName" gorm:"notNull"`
 	Description  string    `json:"dishDescription"`
-	Price        float32   `json:"dishPrice" gorm:"notNull"`
+	Price        float64   `json:"dishPrice" gorm:"notNull"`
 	Quantity     uint      `json:"dishQuantity" gorm:"notNull"`
 	Category     uint      `json:"category" gorm:"type:bigInt;foreignKey:category.id;notNull"`
 	IsVeg        bool      `json:"isVeg" gorm:"default:false"`
